Move Comment and Reply into content folder
