+++ src/main/java/run/halo/app/content/comment/CommentEmailOwner.java
@@ -4,7 +4,7 @@ import java.util.LinkedHashMap;
 import java.util.Map;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.util.Assert;
-import run.halo.app.core.extension.Comment;
+import run.halo.app.core.extension.content.Comment;
 
 /**
  * <p>The creator info of the comment.</p>
