+++ src/main/java/run/halo/app/content/comment/CommentServiceImpl.java
@@ -13,8 +13,8 @@ import org.springframework.security.core.context.ReactiveSecurityContextHolder;
 import org.springframework.stereotype.Component;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
-import run.halo.app.core.extension.Comment;
 import run.halo.app.core.extension.User;
+import run.halo.app.core.extension.content.Comment;
 import run.halo.app.extension.Extension;
 import run.halo.app.extension.ListResult;
 import run.halo.app.extension.ReactiveExtensionClient;
