+++ src/main/java/run/halo/app/core/extension/endpoint/CommentEndpoint.java
@@ -20,8 +20,8 @@ import run.halo.app.content.comment.CommentService;
 import run.halo.app.content.comment.ListedComment;
 import run.halo.app.content.comment.ReplyRequest;
 import run.halo.app.content.comment.ReplyService;
-import run.halo.app.core.extension.Comment;
-import run.halo.app.core.extension.Reply;
+import run.halo.app.core.extension.content.Comment;
+import run.halo.app.core.extension.content.Reply;
 import run.halo.app.extension.ListResult;
 import run.halo.app.extension.router.QueryParamBuildUtil;
 import run.halo.app.infra.utils.HaloUtils;
