+++ src/main/java/run/halo/app/core/extension/reconciler/PostReconciler.java
@@ -15,7 +15,7 @@ import org.springframework.util.Assert;
 import run.halo.app.content.ContentService;
 import run.halo.app.content.PostService;
 import run.halo.app.content.permalinks.PostPermalinkPolicy;
-import run.halo.app.core.extension.Comment;
+import run.halo.app.core.extension.content.Comment;
 import run.halo.app.core.extension.content.Post;
 import run.halo.app.core.extension.content.Snapshot;
 import run.halo.app.event.post.PostPublishedEvent;
