+++ src/main/java/run/halo/app/theme/finders/CommentFinder.java
@@ -2,7 +2,7 @@ package run.halo.app.theme.finders;
 
 import org.springframework.lang.Nullable;
 import reactor.core.publisher.Mono;
-import run.halo.app.core.extension.Comment;
+import run.halo.app.core.extension.content.Comment;
 import run.halo.app.extension.ListResult;
 import run.halo.app.extension.Ref;
 import run.halo.app.theme.finders.vo.CommentVo;
