+++ src/main/java/run/halo/app/theme/finders/vo/CommentVo.java
@@ -6,7 +6,7 @@ import lombok.EqualsAndHashCode;
 import lombok.Value;
 import lombok.With;
 import run.halo.app.content.comment.OwnerInfo;
-import run.halo.app.core.extension.Comment;
+import run.halo.app.core.extension.content.Comment;
 import run.halo.app.extension.MetadataOperator;
 
 /**
