+++ src/main/java/run/halo/app/theme/finders/vo/ReplyVo.java
@@ -7,7 +7,7 @@ import lombok.ToString;
 import lombok.Value;
 import lombok.With;
 import run.halo.app.content.comment.OwnerInfo;
-import run.halo.app.core.extension.Reply;
+import run.halo.app.core.extension.content.Reply;
 import run.halo.app.extension.MetadataOperator;
 
 /**
