+++ src/test/java/run/halo/app/content/comment/CommentServiceImplTest.java
@@ -28,8 +28,8 @@ import org.springframework.util.MultiValueMap;
 import reactor.core.publisher.Mono;
 import reactor.test.StepVerifier;
 import run.halo.app.content.TestPost;
-import run.halo.app.core.extension.Comment;
 import run.halo.app.core.extension.User;
+import run.halo.app.core.extension.content.Comment;
 import run.halo.app.core.extension.content.Post;
 import run.halo.app.extension.ListResult;
 import run.halo.app.extension.Metadata;
