+++ src/test/java/run/halo/app/core/extension/reconciler/CommentReconcilerTest.java
@@ -26,9 +26,9 @@ import org.mockito.ArgumentCaptor;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
-import run.halo.app.core.extension.Comment;
-import run.halo.app.core.extension.Reply;
+import run.halo.app.core.extension.content.Comment;
 import run.halo.app.core.extension.content.Post;
+import run.halo.app.core.extension.content.Reply;
 import run.halo.app.extension.ExtensionClient;
 import run.halo.app.extension.GroupVersionKind;
 import run.halo.app.extension.Metadata;
