+++ src/test/java/run/halo/app/theme/endpoint/CommentFinderEndpointTest.java
@@ -22,8 +22,8 @@ import run.halo.app.content.comment.CommentRequest;
 import run.halo.app.content.comment.CommentService;
 import run.halo.app.content.comment.ReplyRequest;
 import run.halo.app.content.comment.ReplyService;
-import run.halo.app.core.extension.Comment;
-import run.halo.app.core.extension.Reply;
+import run.halo.app.core.extension.content.Comment;
+import run.halo.app.core.extension.content.Reply;
 import run.halo.app.extension.ListResult;
 import run.halo.app.extension.Ref;
 import run.halo.app.theme.finders.CommentFinder;
